+++ src/StashUserscriptLibrary.js
@@ -1,6 +1,6 @@
 // Stash Userscript Library
 // Exports utility functions and a Stash class that emits events whenever a GQL response is received and whenenever a page navigation change is detected
-// version 0.26.3
+// version 0.27.0
 
 (function () {
     'use strict';
@@ -887,7 +887,7 @@
         };
     };
 
-    if (!window.stash) {
-        window.stash = stash();
+    if (!unsafeWindow.stash) {
+        unsafeWindow.stash = stash();
     }
 })();