+++ src/body/Stash Batch Query Edit.user.js
@@ -13,7 +13,7 @@
         createElementFromHTML,
         updateTextInput,
         sortElementChildren,
-    } = window.stash;
+    } = unsafeWindow.stash;
 
     let running = false;
     const buttons = [];