+++ src/body/Stash Batch Result Toggle.user.js
@@ -12,7 +12,7 @@
         createElementFromHTML,
         updateTextInput,
         sortElementChildren,
-    } = window.stash;
+    } = unsafeWindow.stash;
 
     let running = false;
     const buttons = [];