+++ src/header/Stash New Performer Filter Button.user.js
@@ -2,7 +2,7 @@
 // @name        Stash New Performer Filter Button
 // @namespace   %NAMESPACE%
 // @description Adds a button to the performers page to switch to a new performers filter
-// @version     0.2.5
+// @version     0.3.0
 // @author      7dJx1qP
 // @match       %MATCHURL%
 // @grant       unsafeWindow