+++ src/header/Stash Performer Image Cropper.user.js
@@ -2,7 +2,7 @@
 // @name        Stash Performer Image Cropper
 // @namespace   %NAMESPACE%
 // @description Adds an image cropper to performer page
-// @version     0.1.4
+// @version     0.2.0
 // @author      7dJx1qP
 // @match       %MATCHURL%
 // @resource    IMPORTED_CSS https://raw.githubusercontent.com/fengyuanchen/cropperjs/main/dist/cropper.min.css