+++ src/header/Stash Performer Tagger Additions.user.js
@@ -2,7 +2,7 @@
 // @name        Stash Performer Tagger Additions
 // @namespace   %NAMESPACE%
 // @description Adds performer birthdate and url to tagger view. Makes clicking performer name open stash profile in new tab instead of current tab.
-// @version     0.1.5
+// @version     0.2.0
 // @author      7dJx1qP
 // @match       %MATCHURL%
 // @grant       unsafeWindow