+++ src/header/Stash Performer URL Searchbox.user.js
@@ -2,7 +2,7 @@
 // @name        Stash Performer URL Searchbox
 // @namespace   %NAMESPACE%
 // @description Adds a search by performer url textbox to the performers page
-// @version     0.1.5
+// @version     0.2.0
 // @author      7dJx1qP
 // @match       %MATCHURL%
 // @grant       unsafeWindow