+++ src/header/Stash Scene Tagger Additions.user.js
@@ -2,7 +2,7 @@
 // @name        Stash Scene Tagger Additions
 // @namespace   %NAMESPACE%
 // @description Adds scene duration and filepath to tagger view.
-// @version     0.1.2
+// @version     0.2.0
 // @author      7dJx1qP
 // @match       %MATCHURL%
 // @grant       unsafeWindow