+++ src/header/Stash Studio Image And Parent On Create.user.js
@@ -2,7 +2,7 @@
 // @name        Stash Studio Image And Parent On Create
 // @namespace   %NAMESPACE%
 // @description Set studio image and parent when creating from StashDB. Requires userscript_functions stash plugin
-// @version     0.1.5
+// @version     0.2.0
 // @author      7dJx1qP
 // @match       %MATCHURL%
 // @grant       unsafeWindow